Use int instead of size_t in MultiFormatArchive API
Change fileId, fileSizeUncompressed, and related methods to use int
throughout Archive.h/Archive.cpp. Update callers to drop redundant
casts and use -1 / >= 0 for the not-found sentinel.

prompt: change Archive.h Archive.cpp to use int instead of size_t, update callers and simplify int <=> size_t casts, change type of local variables in callers if helpful

Use int instead of size_t in more helpers and AllocArray callers

Change Arena AllocArray/AllocArrayTemp to take int n (matching Base.h).
Update callers to drop redundant size_t/int casts. Also switch
HandleTealDocTag, IsAnnotationTypeInArray, and CliWrite to int, and
simplify related locals in ChmFile, EngineEbook, and EbookDoc.

prompt: change AllocArray to use int for n and update callers, simplify casts if possible; same for HandleTealDocTag and IsAnnotationTypeInArray; commit all

Use int instead of size_t in ByteReader API

Change len, offsets, and Unpack parameters to int in ByteReader.h/cpp.
Update ChmFile, GdiPlus, and ExifDump callers to use int locals and
drop redundant size_t casts.

prompt: change ByteReader to use int instead of size_t; propagate those changes to callers to simplify the code and casts to/from size_t

Use int instead of size_t in ExifDump helpers

Change HEIF/TIFF sniffing, blob copy, and CliWrite to use int for
lengths and offsets; keep size_t casts only at malloc/memcpy/fwrite
boundaries.

prompt: update ExifDump.cpp to use int instead of size_t

Use int instead of size_t in FileHistory API

Change Get/FindByName to use int indices; update Favorites callers
and FileState::index (via gen-settings). Simplify loops and drop
size_t casts in sort/path-check helpers.

prompt: change FileHistory to use int instead of size_t and propagate changes to callers and update them to simplify int <=> size_t casting

Change ByteWriter to use int instead of size_t

Update Zip.cpp and MakeLzSA.cpp callers to use constexpr int size constants and drop unnecessary int casts. Also simplify a length check in WebpReader.cpp.

prompt: change ByteWriter to use int instead of size_t; update callers to simplify int <=> size_t casting

int

+++ src/ChmDump.cpp
@@ -16,12 +16,12 @@
 #include "EbookBase.h"
 #include "ChmFile.h"
 
-static void CliWrite(Str s, size_t n = 0) {
+static void CliWrite(Str s, int n = 0) {
     if (!s) {
         return;
     }
     if (n == 0) {
-        n = (size_t)s.len;
+        n = s.len;
     }
     HANDLE h = GetStdHandle(STD_OUTPUT_HANDLE);
     if (h && h != INVALID_HANDLE_VALUE) {
@@ -29,7 +29,7 @@ static void CliWrite(Str s, size_t n = 0) {
         WriteFile(h, s.s, (DWORD)n, &written, nullptr);
         return;
     }
-    fwrite(s.s, 1, n, stdout);
+    fwrite(s.s, 1, (size_t)n, stdout);
 }
 
 static void CliPrint(Str s) {
