+++ src/MobiDoc.cpp
@@ -738,7 +738,7 @@ void MobiDoc::LoadImages() {
     if (0 == imagesCount) {
         return;
     }
-    images = AllocArray<Str>((int)imagesCount);
+    images = AllocArray<Str>(imagesCount);
 
     for (int i = 0; i < imagesCount; i++) {
         if (!LoadImage(i)) {
